+++ ios/Classes/SwiftFlutterContactsPlugin.swift
@@ -162,17 +162,19 @@ public enum FlutterContacts {
         return groups
     }
 
-    static func fetchGroupMemberships(_ store: CNContactStore, _ groups: [CNGroup]) -> [String: [Int]] {
+    static func fetchGroupMemberships(_ store: CNContactStore, _ groups: [CNGroup], forContactId contactId: String? = nil) -> [String: [Int]] {
         var memberships = [String: [Int]]()
         for (groupIndex, group) in groups.enumerated() {
             let request = CNContactFetchRequest(keysToFetch: [CNContactIdentifierKey] as [CNKeyDescriptor])
             request.predicate = CNContact.predicateForContactsInGroup(withIdentifier: group.identifier)
             do {
                 try store.enumerateContacts(with: request) { (contact, _) -> Void in
-                    if let contactGroups = memberships[contact.identifier] {
-                        memberships[contact.identifier] = contactGroups + [groupIndex]
-                    } else {
-                        memberships[contact.identifier] = [groupIndex]
+                    if contactId == nil || contact.identifier == contactId {
+                        if let contactGroups = memberships[contact.identifier] {
+                            memberships[contact.identifier] = contactGroups + [groupIndex]
+                        } else {
+                            memberships[contact.identifier] = [groupIndex]
+                        }
                     }
                 }
             } catch {
@@ -231,6 +233,7 @@ public enum FlutterContacts {
     // Updates an existing contact in the database.
     static func update(
         _ args: [String: Any?],
+        _ withGroups: Bool,
         _ includeNotesOnIos13AndAbove: Bool
     ) throws -> [String: Any?]? {
         // First, fetch the original contact.
@@ -284,6 +287,27 @@ public enum FlutterContacts {
             let saveRequest = CNSaveRequest()
             saveRequest.update(contact)
             try store.execute(saveRequest)
+
+            // Update group membership
+            if withGroups {
+                let groups = fetchGroups(store)
+                let groupMemberships = fetchGroupMemberships(store, groups, forContactId: contact.identifier)
+                for groupIndex in groupMemberships[contact.identifier] ?? [] {
+                    let deleteRequest = CNSaveRequest()
+                    deleteRequest.removeMember(contact, from: groups[groupIndex])
+                    try store.execute(deleteRequest)
+                }
+
+                let groupIds = Set((args["groups"] as! [[String: Any]]).map { Group(fromMap: $0).id })
+                for group in groups {
+                    if groupIds.contains(group.identifier) {
+                        let addRequest = CNSaveRequest()
+                        addRequest.addMember(contact, to: group)
+                        try store.execute(addRequest)
+                    }
+                }
+            }
+
             return Contact(fromContact: contact).toMap()
         } else {
             return nil
@@ -307,6 +331,65 @@ public enum FlutterContacts {
             saveRequest.delete(contact.mutableCopy() as! CNMutableContact)
         }
         try store.execute(saveRequest)
+    }
+
+    static func getGroups() -> [[String: Any]] {
+        let store = CNContactStore()
+        let groups = fetchGroups(store)
+        return groups.map { Group(fromGroup: $0).toMap() }
+    }
+
+    static func insertGroup(_ args: [String: Any]) throws -> [String: Any] {
+        let group = Group(fromMap: args)
+        let newGroup = CNMutableGroup()
+        newGroup.name = group.name
+
+        let saveRequest = CNSaveRequest()
+        saveRequest.add(newGroup, toContainerWithIdentifier: nil)
+        try CNContactStore().execute(saveRequest)
+
+        return Group(fromGroup: newGroup).toMap()
+    }
+
+    static func updateGroup(_ args: [String: Any]) throws -> [String: Any] {
+        let group = Group(fromMap: args)
+
+        let store = CNContactStore()
+        let groups = fetchGroups(store)
+
+        for g in groups {
+            if g.identifier == group.id {
+                let updatedGroup = g.mutableCopy() as! CNMutableGroup
+                updatedGroup.name = group.name
+
+                let saveRequest = CNSaveRequest()
+                saveRequest.update(updatedGroup)
+                try CNContactStore().execute(saveRequest)
+
+                return Group(fromGroup: updatedGroup).toMap()
+            }
+        }
+
+        return args
+    }
+
+    static func deleteGroup(_ args: [String: Any]) throws {
+        let group = Group(fromMap: args)
+
+        let store = CNContactStore()
+        let groups = fetchGroups(store)
+
+        for g in groups {
+            if g.identifier == group.id {
+                let deletedGroup = g.mutableCopy() as! CNMutableGroup
+
+                let saveRequest = CNSaveRequest()
+                saveRequest.delete(deletedGroup)
+                try CNContactStore().execute(saveRequest)
+
+                return
+            }
+        }
     }
 
     private static func clearFields(
@@ -443,10 +526,11 @@ public class SwiftFlutterContactsPlugin: NSObject, FlutterPlugin, FlutterStreamH
             DispatchQueue.global(qos: .userInteractive).async {
                 let args = call.arguments as! [Any?]
                 let c = args[0] as! [String: Any?]
-                let includeNotesOnIos13AndAbove = args[1] as! Bool
+                let withGroups = args[1] as! Bool
+                let includeNotesOnIos13AndAbove = args[2] as! Bool
                 do {
                     let contact = try FlutterContacts.update(
-                        c, includeNotesOnIos13AndAbove
+                        c, withGroups, includeNotesOnIos13AndAbove
                     )
                     result(contact)
                 } catch {
@@ -461,6 +545,64 @@ public class SwiftFlutterContactsPlugin: NSObject, FlutterPlugin, FlutterStreamH
             DispatchQueue.global(qos: .userInteractive).async {
                 do {
                     try FlutterContacts.delete(call.arguments as! [String])
+                    result(nil)
+                } catch {
+                    result(FlutterError(
+                        code: "unknown error",
+                        message: "unknown error",
+                        details: error.localizedDescription
+                    ))
+                }
+            }
+        case "getGroups":
+            DispatchQueue.global(qos: .userInteractive).async {
+                do {
+                    let groups = try FlutterContacts.getGroups()
+                    result(groups)
+                } catch {
+                    result(FlutterError(
+                        code: "unknown error",
+                        message: "unknown error",
+                        details: error.localizedDescription
+                    ))
+                }
+            }
+        case "insertGroup":
+            DispatchQueue.global(qos: .userInteractive).async {
+                do {
+                    let args = call.arguments as! [Any?]
+                    let g = args[0] as! [String: Any]
+                    let group = try FlutterContacts.insertGroup(g)
+                    result(group)
+                } catch {
+                    result(FlutterError(
+                        code: "unknown error",
+                        message: "unknown error",
+                        details: error.localizedDescription
+                    ))
+                }
+            }
+        case "updateGroup":
+            DispatchQueue.global(qos: .userInteractive).async {
+                do {
+                    let args = call.arguments as! [Any?]
+                    let g = args[0] as! [String: Any]
+                    let group = try FlutterContacts.updateGroup(g)
+                    result(group)
+                } catch {
+                    result(FlutterError(
+                        code: "unknown error",
+                        message: "unknown error",
+                        details: error.localizedDescription
+                    ))
+                }
+            }
+        case "deleteGroup":
+            DispatchQueue.global(qos: .userInteractive).async {
+                do {
+                    let args = call.arguments as! [Any?]
+                    let g = args[0] as! [String: Any]
+                    try FlutterContacts.deleteGroup(g)
                     result(nil)
                 } catch {
                     result(FlutterError(